feat: KatakanaSpellCheckValidatorを実装。 test: KatakanaSpellCheckValidatorTestsを一部のテストケースをJAVA版からベタ移植。グリーンバー。

+++ RedPen.Net.Core/Config/ValidatorConfiguration.cs
@@ -122,7 +122,7 @@ namespace RedPen.Net.Core.Config
 
     public interface IMinRatioConfigParameter
     {
-        public float MinRatio { get; init; }
+        public double MinRatio { get; init; }
     }
 
     public interface IMinFreqConfigParameter
